@@ -10,16 +10,16 @@
     "@emotion/styled": "^11.3.0",
     "@material-ui/core": "^5.0.0-beta.4",
     "@material-ui/icons": "^5.0.0-beta.4",
-    "@mui/material": "^5.0.1",
+    "@mui/material": "^5.0.2",
     "@mui/styles": "^5.0.1",
     "@mui/x-data-grid": "^4.0.0",
     "@visx/xychart": "^2.1.1",
-    "ag-grid-community": "^26.0.0",
-    "ag-grid-react": "^26.0.0",
+    "ag-grid-community": "^26.1.0",
+    "ag-grid-react": "^26.1.0",
     "css-loader": "^6.3.0",
     "html-webpack-plugin": "^5.3.2",
     "mui-plus": "^0.0.15",
-    "playwright": "^1.15.0",
+    "playwright": "^1.15.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-spring": "^9.2.4",
@@ -27,7 +27,7 @@
     "serve-handler": "^6.1.3",
     "style-loader": "^3.3.0",
     "table": "^6.7.1",
-    "webpack": "^4.46.0",
+    "webpack": "^5.56.1",
     "webpack-cli": "^4.8.0"
   }
 }
