@@ -63,26 +63,26 @@
     "@eps1lon/enzyme-adapter-react-17": "^0.1.0",
     "@material-ui/monorepo": "https://github.com/mui-org/material-ui.git#master",
     "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.1",
+    "@mui/material": "^5.0.2",
     "@mui/styles": "^5.0.1",
     "@mui/utils": "^5.0.1",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-node-resolve": "^13.0.5",
     "@rollup/plugin-replace": "^3.0.0",
-    "@testing-library/react": "^12.1.0",
+    "@testing-library/react": "^12.1.2",
     "@types/chai": "^4.2.22",
     "@types/chai-dom": "^0.0.11",
     "@types/enzyme": "^3.10.9",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.10.1",
-    "@types/prettier": "^2.4.0",
-    "@types/react": "^17.0.24",
+    "@types/node": "^16.10.2",
+    "@types/prettier": "^2.4.1",
+    "@types/react": "^17.0.27",
     "@types/react-dom": "^17.0.9",
     "@types/requestidlecallback": "^0.3.4",
     "@types/sinon": "^10.0.4",
     "@types/yargs": "^17.0.3",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
     "argos-cli": "^0.3.3",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-plugin-optimize-clsx": "^2.6.2",
@@ -91,7 +91,7 @@
     "chai": "^4.3.4",
     "chai-dom": "^1.10.0",
     "compression-webpack-plugin": "^9.0.0",
-    "concurrently": "^6.2.1",
+    "concurrently": "^6.3.0",
     "cross-env": "^7.0.3",
     "enzyme": "^3.11.0",
     "eslint": "^7.32.0",
@@ -103,7 +103,7 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-react": "^7.26.0",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
     "format-util": "^1.0.5",
     "glob-gitignore": "^1.0.14",
@@ -117,11 +117,11 @@
     "lerna": "^4.0.0",
     "mocha": "^9.1.2",
     "nyc": "^15.1.0",
-    "playwright": "^1.15.0",
+    "playwright": "^1.15.1",
     "prettier": "^2.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "rollup": "^2.57.0",
+    "rollup": "^2.58.0",
     "rollup-plugin-cleaner": "^1.0.0",
     "rollup-plugin-command": "^1.1.3",
     "rollup-plugin-commonjs": "^10.1.0",
@@ -135,9 +135,9 @@
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-processor-styled-components": "^1.10.0",
-    "typedoc": "^0.22.4",
+    "typedoc": "^0.22.5",
     "typescript": "^4.4.3",
-    "webpack": "^5.54.0",
+    "webpack": "^5.56.1",
     "webpack-cli": "^4.8.0",
     "yargs": "^17.2.1",
     "yarn-deduplicate": "^3.1.0"
