@@ -17,30 +17,30 @@
   },
   "dependencies": {
     "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.1",
+    "@mui/material": "^5.0.2",
     "@mui/styles": "^5.0.1",
     "@mui/x-data-grid": "5.0.0-beta.2",
     "@mui/x-data-grid-generator": "5.0.0-beta.2",
     "@mui/x-data-grid-pro": "5.0.0-beta.2",
     "@mui/x-license-pro": "5.0.0-beta.2",
-    "@storybook/builder-webpack5": "^6.4.0-beta.1",
-    "@storybook/manager-webpack5": "^6.4.0-beta.1",
+    "@storybook/builder-webpack5": "^6.4.0-beta.3",
+    "@storybook/manager-webpack5": "^6.4.0-beta.3",
     "react": "^17.0.2",
     "react-is": "^17.0.2",
-    "rxjs": "^7.3.0",
-    "webpack": "^5.54.0"
+    "rxjs": "^7.3.1",
+    "webpack": "^5.56.1"
   },
   "devDependencies": {
     "@babel/core": "^7.15.5",
-    "@storybook/addon-a11y": "^6.4.0-beta.1",
-    "@storybook/addon-actions": "^6.4.0-beta.1",
+    "@storybook/addon-a11y": "^6.4.0-beta.3",
+    "@storybook/addon-actions": "^6.4.0-beta.3",
     "@storybook/addon-console": "^1.2.3",
-    "@storybook/addon-docs": "^6.4.0-beta.1",
-    "@storybook/addon-essentials": "^6.4.0-beta.1",
-    "@storybook/addon-storysource": "^6.4.0-beta.1",
-    "@storybook/addon-viewport": "^6.4.0-beta.1",
-    "@storybook/react": "^6.4.0-beta.1",
-    "@storybook/theming": "^6.4.0-beta.1",
+    "@storybook/addon-docs": "^6.4.0-beta.3",
+    "@storybook/addon-essentials": "^6.4.0-beta.3",
+    "@storybook/addon-storysource": "^6.4.0-beta.3",
+    "@storybook/addon-viewport": "^6.4.0-beta.3",
+    "@storybook/react": "^6.4.0-beta.3",
+    "@storybook/theming": "^6.4.0-beta.3",
     "babel-loader": "^8.2.2",
     "core-js": "^3.6.5",
     "react-docgen-typescript-loader": "^3.7.2",
