@@ -2,7 +2,7 @@ const path = require('path');
 
 let defaultPresets;
 
-// We release a ES version of Material-UI.
+// We release a ES version of MUI.
 // It's something that matches the latest official supported features of JavaScript.
 // Nothing more (stage-1, etc), nothing less (require, etc).
 if (process.env.BABEL_ENV === 'es') {
