@@ -173,7 +173,7 @@ const rows = [
   },
   {
     id: 3,
-    col1: 'Material-UI',
+    col1: 'MUI',
     col2: 'is Amazing',
     col3: 'Lorem ipsum is a placeholder text commonly used to demonstrate the visual form of a document or a typeface without relying on meaningful content. Lorem ipsum may be used as a placeholder before final copy is available.',
   },
@@ -191,7 +191,7 @@ const rows = [
   },
   {
     id: 6,
-    col1: 'Material-UI',
+    col1: 'MUI',
     col2: 'is Amazing',
     col3: 'Lorem ipsum may be used as a placeholder before final copy is available.',
   },
