@@ -14,7 +14,7 @@ export function SortedAscendingIcon() {
 const rows = [
   {
     id: 1,
-    name: 'Material-UI',
+    name: 'MUI',
     stars: 28000,
   },
   {
