@@ -9,7 +9,7 @@ export default function RowsGrid() {
         rows={
           [
             { id: 1, name: 'React' },
-            { id: 2, name: 'Material-UI' },
+            { id: 2, name: 'MUI' },
           ] as GridRowModel[]
         }
       />
