@@ -4,13 +4,13 @@ import { useLicenseVerifier, LicenseStatus } from '@mui/x-license-pro';
 function getLicenseErrorMessage(licenseStatus: string) {
   switch (licenseStatus) {
     case LicenseStatus.Expired.toString():
-      return 'Material-UI X License Expired';
+      return 'MUI X License Expired';
     case LicenseStatus.Invalid.toString():
-      return 'Material-UI X Invalid License';
+      return 'MUI X Invalid License';
     case LicenseStatus.NotFound.toString():
-      return 'Material-UI X Unlicensed product';
+      return 'MUI X Unlicensed product';
     default:
-      throw new Error('Material-UI: Unhandled license status.');
+      throw new Error('MUI: Unhandled license status.');
   }
 }
 
