@@ -20,24 +20,42 @@ export const useGridColumnMenu = (apiRef: GridApiRef): void => {
 
   const showColumnMenu = React.useCallback(
     (field: string) => {
-      logger.debug('Opening Column Menu');
-      setGridState((state) => ({
-        ...state,
-        columnMenu: { open: true, field },
-      }));
-      apiRef.current.hidePreferences();
-      forceUpdate();
+      const shouldUpdate = setGridState((state) => {
+        if (state.columnMenu.open && state.columnMenu.field === field) {
+          return state;
+        }
+
+        logger.debug('Opening Column Menu');
+        return {
+          ...state,
+          columnMenu: { open: true, field },
+        };
+      });
+
+      if (shouldUpdate) {
+        apiRef.current.hidePreferences();
+        forceUpdate();
+      }
     },
     [apiRef, forceUpdate, logger, setGridState],
   );
 
   const hideColumnMenu = React.useCallback(() => {
-    logger.debug('Hiding Column Menu');
-    setGridState((state) => ({
-      ...state,
-      columnMenu: { ...state.columnMenu, open: false, field: undefined },
-    }));
-    forceUpdate();
+    const shouldUpdate = setGridState((state) => {
+      if (!state.columnMenu.open && state.columnMenu.field === undefined) {
+        return state;
+      }
+
+      logger.debug('Hiding Column Menu');
+      return {
+        ...state,
+        columnMenu: { ...state.columnMenu, open: false, field: undefined },
+      };
+    });
+
+    if (shouldUpdate) {
+      forceUpdate();
+    }
   }, [forceUpdate, logger, setGridState]);
 
   const toggleColumnMenu = React.useCallback(
