@@ -7,7 +7,7 @@ export function useGridApiContext() {
   if (apiRef === undefined) {
     throw new Error(
       [
-        'Material-UI X: Could not find the data grid context.',
+        'MUI: Could not find the data grid context.',
         'It looks like you rendered your component outside of a DataGrid or DataGridPro parent component.',
         'This can also happen if you are bundling multiple versions of the data grid.',
       ].join('\n'),
