@@ -97,7 +97,7 @@ describe('<DataGrid /> - Layout & Warnings', () => {
     });
 
     // Adaptation of describeConformance()
-    describe('Material-UI component API', () => {
+    describe('MUI component API', () => {
       it(`attaches the ref`, () => {
         const ref = React.createRef<HTMLDivElement>();
         const { container } = render(
@@ -188,9 +188,7 @@ describe('<DataGrid /> - Layout & Warnings', () => {
           // Use timeout to allow simpler tests in JSDOM.
           clock.tick(0);
           // @ts-expect-error need to migrate helpers to TypeScript
-        }).toWarnDev(
-          'Material-UI: useResizeContainer - The parent of the grid has an empty height.',
-        );
+        }).toWarnDev('MUI: useResizeContainer - The parent of the grid has an empty height.');
       });
 
       it('should warn if the container has no intrinsic width', () => {
@@ -205,9 +203,7 @@ describe('<DataGrid /> - Layout & Warnings', () => {
           // Use timeout to allow simpler tests in JSDOM.
           clock.tick(0);
           // @ts-expect-error need to migrate helpers to TypeScript
-        }).toWarnDev(
-          'Material-UI: useResizeContainer - The parent of the grid has an empty width.',
-        );
+        }).toWarnDev('MUI: useResizeContainer - The parent of the grid has an empty width.');
       });
 
       it('should warn when GridCellParams.valueGetter is called with a missing column', () => {
@@ -659,7 +655,7 @@ describe('<DataGrid /> - Layout & Warnings', () => {
           </div>,
         );
         // @ts-expect-error need to migrate helpers to TypeScript
-      }).toErrorDev('Material-UI: `<DataGrid pagination={false} />` is not a valid prop.');
+      }).toErrorDev('MUI: `<DataGrid pagination={false} />` is not a valid prop.');
     });
 
     it('should throw if the rows has no id', function test() {
