@@ -95,7 +95,7 @@ DataGridProRaw.propTypes = {
   checkboxSelectionVisibleOnly: chainPropTypes(PropTypes.bool, (props: any) => {
     if (!props.pagination && props.checkboxSelectionVisibleOnly) {
       return new Error(
-        'Material-UI: The `checkboxSelectionVisibleOnly` prop has no effect when the pagination is not enabled.',
+        'MUI: The `checkboxSelectionVisibleOnly` prop has no effect when the pagination is not enabled.',
       );
     }
     return null;
@@ -266,7 +266,7 @@ DataGridProRaw.propTypes = {
   hideFooterRowCount: chainPropTypes(PropTypes.bool, (props: any) => {
     if (props.pagination && props.hideFooterRowCount) {
       return new Error(
-        'Material-UI: The `hideFooterRowCount` prop has no effect when the pagination is enabled.',
+        'MUI: The `hideFooterRowCount` prop has no effect when the pagination is enabled.',
       );
     }
     return null;
