@@ -2,7 +2,7 @@ import { create } from '@storybook/theming/create';
 
 export default create({
   base: 'light',
-  brandTitle: 'Material-UI X Grid storybook',
-  brandUrl: 'https://material-ui.com/',
-  brandImage: 'https://material-ui.com/static/logo_raw.svg', // TODO replace with a smaller logo
+  brandTitle: 'MUI X Grid storybook',
+  brandUrl: 'https://mui.com/',
+  brandImage: 'https://mui.com/static/logo_raw.svg', // TODO replace with a smaller logo
 });
