@@ -13,9 +13,7 @@ const expiryReg = /^.*EXPIRY=([0-9]+),.*$/;
 
 export function verifyLicense(releaseInfo: string, encodedLicense: string) {
   if (!releaseInfo) {
-    throw new Error(
-      'Material-UI: The release information is missing. Not able to validate license.',
-    );
+    throw new Error('MUI: The release information is missing. Not able to validate license.');
   }
 
   if (!encodedLicense) {
@@ -44,9 +42,7 @@ export function verifyLicense(releaseInfo: string, encodedLicense: string) {
 
   const pkgTimestamp = parseInt(base64Decode(releaseInfo), 10);
   if (Number.isNaN(pkgTimestamp)) {
-    throw new Error(
-      'Material-UI: The release information is invalid. Not able to validate license.',
-    );
+    throw new Error('MUI: The release information is invalid. Not able to validate license.');
   }
 
   if (expiryTimestamp < pkgTimestamp) {
